+++ packages/yourloops/components/patient/remove-patient-dialog.tsx
@@ -43,6 +43,7 @@ import MedicalServiceIcon from '../icons/medical-service-icon'
 import useRemovePatientDialog from './remove-patient-dialog.hook'
 import { type Patient } from '../../lib/patient/models/patient.model'
 import { LoadingButton } from '@mui/lab'
+import TeamUtils from '../../lib/team/team.util'
 
 interface RemovePatientDialogProps {
   patient: Patient | null
@@ -95,7 +96,7 @@ const RemovePatientDialog: FunctionComponent<RemovePatientDialogProps> = ({ onCl
                 key={index}
                 data-testid={`select-option-${team.name}`}
               >
-                {team.code === 'private'
+                {TeamUtils.isPrivate(team)
                   ? <Box display="flex" alignItems="center">
                     <React.Fragment>
                       <Box display="flex" ml={0} mr={1}>