+++ CHANGELOG.md
@@ -7,6 +7,9 @@ It is based on Tidepool Blip 1.27.
 - YLP-1503 Logout idle user after 30 mn of inactivity
 - YLP-2140 As a HCP, I can only edit my medical records
 - YLP-2175 Implement new design system
+- YLP-2192 Team scope dropdown for HCP
+- YLP-2194 Navigation tabs for HCP ("Patients", "Care teams")
+- YLP-2220 Remove "private practice" filter from patients list
 - YLP-2231 Implement new medical reports design for patients
 
 ### Engineering use